Structure the project
Start structuring the project. Please note that my usual TDD approach is hard to perform on Android given a lack of pre-defined APIs. What I would normally do on a mobile platform is setup a BDD approach. This however is not feasible in the scope of this short time project.